Fix bug: Medium Editor initiates before existing post content is available.
When loading an existing post into the editor, currently we wait for
the post subscription to be ready. However, we are not waiting for
the subscription before running `BlogEditor.make`, which sets up
Medium Editor and medium-insert. Since those are both dependent on the
existing post content to be populated into the editor, this was
causing some bugs, the main one being that images would sometimes
disappear when switching between the visual and HTML editor.

By nesting `BlogEditor.make` in the autorun that waits for the
subscription, we can ensure the content is poulated when setting up
the editor.